Run nghttpx in foreground
nghttpx 1.14.0 can reload configuration with SIGHUP.  Using this
feature, reloading configuration is now much easier.  We don't have to
make nghttpx run daemon anymore, because we just send SIGHUP to make
it reload configuration.  To avoid truncation of configuration file in
first reload, we use a common trick to first write to temporary file
and move it to the final configuration file position.

We also fixed the issue that initial default.conf does not have
backend configuration.  We mark the port 80 listener as healthmon for
now.