Add default-tls-secret flag to enable TLS by default
default-tls-secret flag is added to enable TLS by default.  You could
the similar effect by creating Ingress resource for default backend.
But the namespace of Ingress resource that Ingress controller watches,
and that of default backend could differ, and in this case we cannot
use that method.

Some refactoring is done to get certificates from Secret.