Use single sync key, and apply throttling on sync function
Previously, we apply throttling on CheckAndReload function.  It means
that we most likely use older configuration.  In this commit, we apply
throttling on sync function, and use latest configuration if possible.

Since the controller always uses all available information, it is not
necessary to enqueue the object key which has some events.  Therefore,
we use the same key for all events.