Get address from EndpointSlice when processing Service without selectors
If EndpointSlice is enabled, which is default, get addresses from
EndpointSlice when processing Service without selectors.

This removes the need to cache Endpoints when EndpointSlice is
enabled.  This change also removes the watch filter by
discoveryv1.LabelManagedBy for EndpointSlice to cache EndpointSlices
which are not managed by endpointslice controller.  But that label
selector is simply unnecessary for the general case because we just
need EndpointSlices whatever manages it.