Delete asset files by checking its modification time
Previously, there is a race condition that asset files being used are
deleted if backend is reloaded too quickly.  This commit changes the
way of deleting asset files by checking its modification time, and if
it is too old, the file is deleted.  New flag, stale-assets-threshold,
is added to control this threshold.