Retry getting configRevision before reloading configuration
Getting the current configRevision before reloading configuration
might fail if it is performed before nghttpx process gets started and
listens on the API endpoint.  If that happens the configuration inside
Pod becomes inconsistent state because files are not written out.
This change fixes this issue by moving configRevision acquisition
after writing files, and retrying it on failure.